+++ CHANGELOG.md
@@ -22,6 +22,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add reminder notifications to prompt daily weight logging
 - Add option to preview interpolation on own data
 - Add changelog to app
+- Add tooltip when scrolling the linechart
 
 ### Other changes:
 - Using the latest flutter 3.41